@@ -10,15 +10,18 @@
         "slack": "https://codeigniterchat.slack.com"
     },
     "require": {
-        "php": "^7.4 || ^8.0",
-        "codeigniter4/framework": "^4.0",
-        "fakerphp/faker": "^1.23"
+        "php": "^8.1",
+        "codeigniter4/framework": "^4.0"
     },
     "require-dev": {
+        "fakerphp/faker": "^1.9",
         "mikey179/vfsstream": "^1.6",
-        "phpunit/phpunit": "^9.1"
+        "phpunit/phpunit": "^10.5.16"
     },
     "autoload": {
+        "psr-4": {
+            "App\\": "app/"
+        },
         "exclude-from-classmap": [
             "**/Database/Migrations/**"
         ]
