@@ -10,7 +10,7 @@
  |
  | If you set 'display_errors' to '1', CI4's detailed error report will show.
  */
-error_reporting(-1);
+error_reporting(E_ALL);
 ini_set('display_errors', '1');
 
 /*
