@@ -2,7 +2,6 @@
 
 namespace Config;
 
-use CodeIgniter\Config\BaseConfig;
 use Kint\Parser\ConstructablePluginInterface;
 use Kint\Renderer\AbstractRenderer;
 use Kint\Renderer\Rich\TabPluginInterface;
@@ -18,7 +17,7 @@ use Kint\Renderer\Rich\ValuePluginInterface;
  *
  * @see https://kint-php.github.io/kint/ for details on these settings.
  */
-class Kint extends BaseConfig
+class Kint
 {
     /*
     |--------------------------------------------------------------------------
@@ -27,8 +26,7 @@ class Kint extends BaseConfig
     */
 
     /**
-     * @var array<int, ConstructablePluginInterface|string>
-     * @phpstan-var list<class-string<ConstructablePluginInterface>|ConstructablePluginInterface>
+     * @var list<class-string<ConstructablePluginInterface>|ConstructablePluginInterface>|null
      */
     public $plugins;
 
@@ -46,14 +44,12 @@ class Kint extends BaseConfig
     public int $richSort     = AbstractRenderer::SORT_FULL;
 
     /**
-     * @var array<string, string>
-     * @phpstan-var array<string, class-string<ValuePluginInterface>>
+     * @var array<string, class-string<ValuePluginInterface>>|null
      */
     public $richObjectPlugins;
 
     /**
-     * @var array<string, string>
-     * @phpstan-var array<string, class-string<TabPluginInterface>>
+     * @var array<string, class-string<TabPluginInterface>>|null
      */
     public $richTabPlugins;
 
