@@ -58,7 +58,7 @@ class Modules extends BaseModules
      *       ],
      *   ]
      *
-     * @var array
+     * @var array{only?: list<string>, exclude?: list<string>}
      */
     public $composerPackages = [];
 
@@ -72,7 +72,7 @@ class Modules extends BaseModules
      *
      * If it is not listed, only the base application elements will be used.
      *
-     * @var string[]
+     * @var list<string>
      */
     public $aliases = [
         'events',
