@@ -19,17 +19,21 @@ use CodeIgniter\Config\Routing as BaseRouting;
 class Routing extends BaseRouting
 {
     /**
+     * For Defined Routes.
      * An array of files that contain route definitions.
      * Route files are read in order, with the first match
      * found taking precedence.
      *
      * Default: APPPATH . 'Config/Routes.php'
+     *
+     * @var list<string>
      */
     public array $routeFiles = [
         APPPATH . 'Config/Routes.php',
     ];
 
     /**
+     * For Defined Routes and Auto Routing.
      * The default namespace to use for Controllers when no other
      * namespace has been specified.
      *
@@ -38,6 +42,7 @@ class Routing extends BaseRouting
     public string $defaultNamespace = 'App\Controllers';
 
     /**
+     * For Auto Routing.
      * The default controller to use when no other controller has been
      * specified.
      *
@@ -46,6 +51,7 @@ class Routing extends BaseRouting
     public string $defaultController = 'Home';
 
     /**
+     * For Defined Routes and Auto Routing.
      * The default method to call on the controller when no other
      * method has been set in the route.
      *
@@ -54,7 +60,8 @@ class Routing extends BaseRouting
     public string $defaultMethod = 'index';
 
     /**
-     * Whether to translate dashes in URIs to underscores.
+     * For Auto Routing.
+     * Whether to translate dashes in URIs for controller/method to underscores.
      * Primarily useful when using the auto-routing.
      *
      * Default: false
@@ -63,13 +70,12 @@ class Routing extends BaseRouting
 
     /**
      * Sets the class/method that should be called if routing doesn't
-     * find a match. It can be either a closure or the controller/method
-     * name exactly like a route is defined: Users::index
+     * find a match. It can be the controller/method name like: Users::index
      *
      * This setting is passed to the Router class and handled there.
      *
      * If you want to use a closure, you will have to set it in the
-     * class constructor or the routes file by calling:
+     * routes file by calling:
      *
      * $routes->set404Override(function() {
      *    // Do something here
@@ -91,6 +97,7 @@ class Routing extends BaseRouting
     public bool $autoRoute = false;
 
     /**
+     * For Defined Routes.
      * If TRUE, will enable the use of the 'prioritize' option
      * when defining routes.
      *
@@ -99,7 +106,16 @@ class Routing extends BaseRouting
     public bool $prioritize = false;
 
     /**
-     * Map of URI segments and namespaces. For Auto Routing (Improved).
+     * For Defined Routes.
+     * If TRUE, matched multiple URI segments will be passed as one parameter.
+     *
+     * Default: false
+     */
+    public bool $multipleSegmentsOneParam = false;
+
+    /**
+     * For Auto Routing (Improved).
+     * Map of URI segments and namespaces.
      *
      * The key is the first URI segment. The value is the controller namespace.
      * E.g.,
@@ -107,7 +123,18 @@ class Routing extends BaseRouting
      *       'blog' => 'Acme\Blog\Controllers',
      *   ]
      *
-     * @var array [ uri_segment => namespace ]
+     * @var array<string, string>
      */
     public array $moduleRoutes = [];
+
+    /**
+     * For Auto Routing (Improved).
+     * Whether to translate dashes in URIs for controller/method to CamelCase.
+     * E.g., blog-controller -> BlogController
+     *
+     * If you enable this, $translateURIDashes is ignored.
+     *
+     * Default: false
+     */
+    public bool $translateUriToCamelCase = false;
 }
